# backend/openapi.yaml
openapi: 3.1.0
info:
  title: WIT Sprint OS API
  version: "1.0.0"
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    REST API for the WIT Sprint OS delivery hierarchy.

    ## Naming

    The URL vocabulary predates the UI's, and the two do not match. Read it as:

    | URL / payload | The UI calls it |
    | ------------- | --------------- |
    | `product`     | **Module**      |
    | `module`      | **Component**   |

    The hierarchy is `client → project → product → module → sprint → task`,
    which the UI presents as Client → Project → Module → Component → Sprint.

    ## Conventions

    - **PATCH is partial.** Only the keys present in the body are written; an
      absent key leaves its column untouched. This is what makes two concurrent
      edits to different fields both survive. `PATCH /products/{id}/current-sprint`
      is the single exception where an explicit `null` differs from an absent
      key — see that operation.
    - **Ids may be supplied or generated.** Send `id` on create to choose it;
      omit it and the server mints one. A duplicate answers `409`.
    - **Lineage is enforced.** A child may only reference a parent inside its own
      product: a sprint's component, a task's backlog item, a report's sprint. A
      mismatch is a `400`, not a silent link.
    - **Timestamps** are RFC 3339 with an offset (`2026-07-23T02:35:08.59+07:00`).
...
      RFC 3339.
...

    ## Authentication

...
    it stands.

...
security: []
servers:
...
paths:
...
  /products:
    get:
      tags: [products]
      summary: List products (UI "Modules")
      description: Filter by `projectId` or `clientId`; without either, lists every product.
      operationId: listProducts
      parameters:
        - name: projectId
          in: query
          schema: { type: string }
          description: Only products of this project.
        - name: clientId
          in: query
          schema: { type: string }
          description: Only products of this client. Ignored when `projectId` is given.
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: A page of products.
          headers:
            X-Has-More: { $ref: '#/components/headers/XHasMore' }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Product' }
        '503': { $ref: '#/components/responses/Unavailable' }
    post:
      tags: [products]
      summary: Create a product
      description: |
        `clientId` is **not accepted** — it is derived from the project, so a
        product can never be attributed to a client that does not own it.
        `currentSprintId` is not accepted either: a new product has no sprints.
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductCreate' }
      responses:
        '201':
          description: The created product.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '400':
          description: '`projectId` is missing or does not exist.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409': { $ref: '#/components/responses/Conflict' }
  /products/{productId}:
    parameters:
      - $ref: '#/components/parameters/ProductId'
    get:
      tags: [products]
      summary: Get a product
      operationId: getProduct
      responses:
        '200':
          description: The product.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [products]
      summary: Update a product
      description: |
        Partial. `clientId` is not accepted; move the product by changing
        `projectId`, and its client follows.
      operationId: updateProduct
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductUpdate' }
      responses:
        '200':
          description: The updated product.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [products]
      summary: Delete a product
      description: Cascades to its components, sprints, backlog, tasks, decisions and reports.
      operationId: deleteProduct
      responses:
        '204': { description: Deleted. }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /products/{productId}/current-sprint:
    parameters:
      - $ref: '#/components/parameters/ProductId'
    patch:
      tags: [products, sprints]
      summary: Point a product at its current sprint
      description: |
        The one endpoint where an explicit `null` and an absent key differ:

        | Body | Effect |
        | ---- | ------ |
        | `{"currentSprintId": "spr-..."}` | Sets the pointer. The sprint must belong to this product. |
        | `{"currentSprintId": null}` | Clears the pointer. |
        | `{}` | Nothing. Returns the product unchanged. |
      operationId: setProductCurrentSprint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentSprintId:
                  type: [string, 'null']
                  description: A sprint of this product, or null to clear.
      responses:
        '200':
          description: The product.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '400':
          description: The sprint is not a sprint of this product.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404': { $ref: '#/components/responses/NotFound' }
  /products/{productId}/modules:
    parameters:
      - $ref: '#/components/parameters/ProductId'
    get:
      tags: [modules]
      summary: List a product's components
      operationId: listModulesByProduct
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: A page of components, in `position` order.
          headers:
            X-Has-More: { $ref: '#/components/headers/XHasMore' }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Module' }
        '503': { $ref: '#/components/responses/Unavailable' }
    post:
      tags: [modules]
      summary: Create a component
      description: |
        Omit `position` and the server appends it under a row lock, so a burst of
        concurrent creates lands on 0, 1, 2… rather than all on 0.
      operationId: createModule
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ModuleCreate' }
      responses:
        '201':
          description: The created component.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Module' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }

  # ---------------------------------------------- modules (UI "Components") ---
  /modules/{moduleId}:
    parameters:
      - $ref: '#/components/parameters/ModuleId'
    get:
      tags: [modules]
      summary: Get a component
      operationId: getModule
      responses:
        '200':
          description: The component.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Module' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [modules]
      summary: Update a component
      operationId: updateModule
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ModuleUpdate' }
      responses:
        '200':
          description: The updated component.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Module' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [modules]
      summary: Delete a component
      description: |
        Sprints and backlog items of this component survive — their `moduleId`
        is set to null. Only the component itself goes.
      operationId: deleteModule
      responses:
        '204': { description: Deleted. }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /modules/{moduleId}/status:
    parameters:
      - $ref: '#/components/parameters/ModuleId'
    patch:
      tags: [modules]
      summary: Set a component's status
      operationId: updateModuleStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { $ref: '#/components/schemas/ModuleStatus' }
      responses:
        '200':
          description: The updated component.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Module' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  parameters:
    Limit:
      name: limit
      in: query
      description: Page size. Clamped to 1000; anything unparseable falls back to the default.
      schema: { type: integer, minimum: 1, maximum: 1000, default: 200 }
    Offset:
      name: offset
      in: query
      description: |
        Rows to skip. Postgres walks and discards them, so deep offsets get
        progressively more expensive — prefer narrowing by parent id. Values
        above 1,000,000 are refused with a 400 rather than silently answered
        with a different page.
      schema: { type: integer, minimum: 0, maximum: 1000000, default: 0 }
...
    ProductId:
      name: productId
      in: path
      required: true
      description: 'A product id — the UI calls this a Module.'
      schema: { type: string }
    ModuleId:
      name: moduleId
      in: path
      required: true
      description: 'A module id — the UI calls this a Component.'
      schema: { type: string }
...
  headers:
    XHasMore:
      description: Present and `true` when another page exists.
      schema: { type: string, enum: ['true'] }
  responses:
    BadRequest:
      description: The request is malformed, or a value is out of range.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: No such record.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: |
        A uniqueness clash, or a row was busy and the write should be retried.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    TooLarge:
      description: The body is over 1 MB.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unavailable:
      description: |
        Retryable. Either the request took too long — the statement timeout is
        5 s and the whole request is capped at 15 s — or the database could not
        take it: out of connection slots, restarting, or the connection dropped.
        Never a bug on the caller's side, and always worth retrying.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, examples: ['name is required'] }

    # ------------------------------------------------------------- enums ---
...
    Risk:              { type: string, enum: [low, medium, high] }
...
    ProductStatus:     { type: string, enum: [discovery, development, release, maintenance] }
    ModuleStatus:      { type: string, enum: [planned, in-progress, done] }
...
    Timestamp:
      type: string
      format: date-time
      description: RFC 3339 with an offset.
      examples: ['2026-07-23T02:35:08.590146+07:00']
...
    Json:
      type: [object, 'null']
      description: Free-form JSON, capped at 64 KB once decoded.

    # ----------------------------------------------------------- clients ---
...
    Product:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        clientId: { type: string, description: Derived from the project; never set directly. }
        name: { type: string }
        goal: { type: string }
        owner: { type: string }
        deliveryLead: { type: string }
        status: { $ref: '#/components/schemas/ProductStatus' }
        health: { type: integer, minimum: 0, maximum: 100 }
        risk: { $ref: '#/components/schemas/Risk' }
        velocity: { type: integer, minimum: 0 }
        blockedCount: { type: integer, minimum: 0 }
        currentSprintId: { type: [string, 'null'] }
        aiInsight: { $ref: '#/components/schemas/Json' }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
        updatedAt: { $ref: '#/components/schemas/Timestamp' }
    ProductCreate:
      type: object
      required: [projectId, name]
      description: '`clientId` and `currentSprintId` are not accepted here.'
      properties:
        id: { type: string }
        projectId: { type: string }
        name: { type: string }
        goal: { type: string }
        owner: { type: string }
        deliveryLead: { type: string }
        status: { $ref: '#/components/schemas/ProductStatus' }
        health: { type: integer, minimum: 0, maximum: 100, default: 100 }
        risk: { $ref: '#/components/schemas/Risk' }
        velocity: { type: integer, minimum: 0 }
        blockedCount: { type: integer, minimum: 0 }
        aiInsight: { $ref: '#/components/schemas/Json' }
    ProductUpdate:
      type: object
      description: '`clientId` is not accepted; change `projectId` to move the product.'
      properties:
        projectId: { type: string }
        name: { type: string }
        goal: { type: string }
        owner: { type: string }
        deliveryLead: { type: string }
        status: { $ref: '#/components/schemas/ProductStatus' }
        health: { type: integer, minimum: 0, maximum: 100 }
        risk: { $ref: '#/components/schemas/Risk' }
        velocity: { type: integer, minimum: 0 }
        blockedCount: { type: integer, minimum: 0 }
        aiInsight: { $ref: '#/components/schemas/Json' }

    # ---------------------------------------- modules (UI "Components") ---
    Module:
      type: object
      properties:
        id: { type: string }
        productId: { type: string }
        name: { type: string }
        owner: { type: string }
        status: { $ref: '#/components/schemas/ModuleStatus' }
        position: { type: integer, minimum: 0 }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
        updatedAt: { $ref: '#/components/schemas/Timestamp' }
    ModuleCreate:
      type: object
      required: [name]
      properties:
        id: { type: string }
        name: { type: string }
        owner: { type: string, default: Unassigned }
        status: { $ref: '#/components/schemas/ModuleStatus' }
        position: { type: integer, minimum: 0, description: Omit to append. }
    ModuleUpdate:
      type: object
      properties:
        name: { type: string }
        owner: { type: string }
        status: { $ref: '#/components/schemas/ModuleStatus' }
        position: { type: integer, minimum: 0 }

    # ----------------------------------------------------------- sprints ---
...